Extract BPF support code
Move the generic BPF event pollers as well as the facilities to handle
BPF related structures such as the objects, maps, etc to their own files

Test Plan
=========

CI
